Improved ASIN support

+++ support/management/commands/webmunk_fetch_amazon_asin_items.py
@@ -6,6 +6,7 @@ import re
 
 from django.core.management.base import BaseCommand
 from django.db.models import Q
+from django.utils import timezone
 
 from passive_data_kit.models import DataPoint
 
@@ -17,10 +18,10 @@ class Command(BaseCommand):
     def add_arguments(self, parser):
         pass
 
-    def handle(self, *args, **options): # pylint: disable=too-many-branches
+    def handle(self, *args, **options): # pylint: disable=too-many-branches, too-many-locals
         query = None
 
-        latest_asin_item = AmazonASINItem.objects.all().order_by('-added').first()
+        latest_asin_item = AmazonASINItem.objects.all().order_by('-pk').first()
 
         seconds_window = 0
 
@@ -28,11 +29,14 @@ class Command(BaseCommand):
             while query is None or DataPoint.objects.filter(query).count() == 0:
                 seconds_window = seconds_window + 60
 
+                window_end = latest_asin_item.added + datetime.timedelta(seconds=seconds_window)
+
                 query = Q(secondary_identifier='webmunk-asin-item')
                 query = query & Q(created__gt=latest_asin_item.added)
-                query = query & Q(created__lte=latest_asin_item.added + datetime.timedelta(seconds=seconds_window))
+                query = query & Q(created__lte=window_end)
 
-            print('SINCE: %s - %d - %d' % (latest_asin_item.added.isoformat(), DataPoint.objects.filter(query).count(), seconds_window))
+                if window_end > timezone.now():
+                    break
         else:
             query = Q(secondary_identifier='webmunk-asin-item').order_by('created')[:250]
 
@@ -43,7 +47,7 @@ class Command(BaseCommand):
 
             props = point.fetch_properties()
 
-            page_url = props.get('url!', None)
+            page_url = props.get('url!', props.get('url*', None))
 
             if page_url is not None and '/dp/' in page_url:
                 matches = re.findall('.*/dp/(.+?)/.*', page_url)
@@ -54,7 +58,7 @@ class Command(BaseCommand):
 
                         asins.append(matched_item)
 
-            element = props.get('element-content!', None)
+            element = props.get('element-content!', props.get('element-content*', None))
 
             if element is not None:
                 matches = re.findall('.*/dp/(.+?)/.*', element)
@@ -72,12 +76,20 @@ class Command(BaseCommand):
                         asins.append(matched_item)
 
             for asin in asins:
+                if '"' in asin:
+                    tokens = asin.split('"')
+
+                    asin = tokens[0]
+
                 if AmazonASINItem.objects.filter(asin=asin).count() == 0:
                     AmazonASINItem.objects.create(asin=asin, added=point.created, updated=point.created)
 
             if last_created is None or point.created > last_created:
                 last_created = point.created
 
-        AmazonASINItem.objects.filter(asin__startswith='20', asin__contains='T').filter(asin__contains=':').filter(asin__contains='+').delete()
+        if last_created is not None:
+            print('SINCE: %s - %d - %d' % (latest_asin_item.added.isoformat(), DataPoint.objects.filter(query).count(), seconds_window))
 
-        AmazonASINItem.objects.create(asin=last_created.isoformat(), added=last_created, updated=last_created)
+            AmazonASINItem.objects.filter(asin__startswith='20', asin__contains='T').filter(asin__contains=':').filter(asin__contains='+').delete()
+
+            AmazonASINItem.objects.create(asin=last_created.isoformat(), added=last_created, updated=last_created)
